+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.8]
+
+### Changed
+
+- 增加客户端调用时的网络错误日志
+- ThriftClientContext 增加beanName常量, 用于bean依赖排序
+    - `@DependsOn(ThriftClientContext.BEAN_NAME)` 可以在bean初始化过程中提前进行thrift调用
+
 ## [1.0.7]
 
 ### Added